Distribute methods between classes
